Update controllers.py
1. I fixed bug in line 115 "image_base64 = base64.b64decode(custom_logo)"
2. I maked logo as URL, instread base64
3. I maked separate website custom logo on url '/logo'
and internal classic odoo logo on url '/web/binary/company_logo'

4. but website logo become static
5. each request to website will requested website logo. And on each request website logo will resizing(look at code with 'image_resize_image'). Then resized logo convert to base64(look at code 'b64decode'). Then base64 converted to bites(look at 'ByteIO'). Then file with send to client.
This algorithm is incorrect!!! This is very slowly and make CPU with to converting logo at each request for each client. We use CDN for static files but this algorithm are make our site slowly even with Content Delivery Network.
Also on cloud VPS we are paid for CPU using. But this with this algorithm we have to paid money for converting logo!